split the general and tutoring events into their own files
